Public Documentation

Documentation for JunctionTrees.jl's public interface.

Index

JunctionTrees

JunctionTrees.BackwardPass

JunctionTrees.ForwardPass

JunctionTrees.JointMarginals

JunctionTrees.Marginals

JunctionTrees.UnnormalizedMarginals

JunctionTrees.Factor

JunctionTrees.LastStage

JunctionTrees.compile_algo

JunctionTrees.marg

JunctionTrees.norm

JunctionTrees.product

JunctionTrees.redu

Public Interface

JunctionTrees — Module

Main module for JunctionTrees.jl – a Julia implementation of the junction tree algorithm.

One main function is exported from this module for public use:

compile_algo. Compiles and returns an expression that computes the posterior marginals of the model given evidence using the junction tree algorithm.

Exports

BackwardPass

Factor

ForwardPass

JointMarginals

LastStage

Marginals

UnnormalizedMarginals

compile_algo

marg

norm

product

redu

JunctionTrees.compile_algo — Function

compile_algo(uai_filepath::String; uai_evid_filepath, td_filepath, apply_partial_evaluation, last_stage, smart_root_selection) -> Expr

Return an expression of the junction tree algorithm that extracts the marginals of all the variables in the model.

Arguments

uai_filepath::String: path to the model file defined in the UAI model file format.

uai_evid_filepath::String = "": path to the evidence file defined in the UAI evidence file format.

td_filepath::String = "": path to a pre-constructed junction tree defined in the PACE graph format.

apply_partial_evaluation::Bool = false: optimize the algorithm using partial evaluation.

last_stage::LastStage = Marginals: return an expression up to the given stage. The options are ForwardPass, BackwardPass, JointMarginals, UnnormalizedMarginals and Marginals.

smart_root_selection::Bool = select as root the cluster with the largest state space.

Examples

algo = compile_algo("../examples/problems/paskin/paskin.uai")
eval(algo)
obsvars, obsvals = Int64[], Int64[]
marginals = run_algo(obsvars, obsvals)

# output

6-element Vector{Factor{Float64, 1}}:
 Factor{Float64, 1}((1,), [0.33480077287635474, 0.33039845424729053, 0.33480077287635474])
 Factor{Float64, 1}((2,), [0.378700415763991, 0.621299584236009])
 Factor{Float64, 1}((3,), [0.3632859624875086, 0.6367140375124913])
 Factor{Float64, 1}((4,), [0.6200692707149191, 0.37993072928508087])
 Factor{Float64, 1}((5,), [0.649200314859223, 0.350799685140777])
 Factor{Float64, 1}((6,), [0.5968155611613972, 0.4031844388386027])

algo = compile_algo(
         "../examples/problems/paskin/paskin.uai",
         uai_evid_filepath = "../examples/problems/paskin/paskin.uai.evid")
eval(algo)
obsvars, obsvals = JunctionTrees.read_uai_evid_file("../examples/problems/paskin/paskin.uai.evid")
marginals = run_algo(obsvars, obsvals)

# output

6-element Vector{Factor{Float64, 1}}:
 Factor{Float64, 1}((1,), [1.0, 0.0, 0.0])
 Factor{Float64, 1}((2,), [0.0959432982733719, 0.9040567017266281])
 Factor{Float64, 1}((3,), [0.07863089300137578, 0.9213691069986242])
 Factor{Float64, 1}((4,), [0.8440129077674895, 0.15598709223251056])
 Factor{Float64, 1}((5,), [0.9015456486772953, 0.09845435132270475])
 Factor{Float64, 1}((6,), [0.6118571666785584, 0.3881428333214415])

JunctionTrees.Factor — Type

struct Factor{T, N}

Fields

vars

vals

Encodes a discrete function over the set of variables vars that maps each instatiation of vars into a nonnegative number in vals.

JunctionTrees.product — Function

product(A::Factor{T}, B::Factor{T}) -> Factor

Compute a factor product of tables A and B.

Examples

A = Factor{Float64,2}((2, 3), [0.5 0.7; 0.1 0.2])
B = Factor{Float64,2}((1, 2), [0.5 0.8; 0.1 0.0; 0.3 0.9])
product(A, B)

# output

Factor{Float64, 3}((1, 2, 3), [0.25 0.08000000000000002; 0.05 0.0; 0.15 0.09000000000000001;;; 0.35 0.16000000000000003; 0.06999999999999999 0.0; 0.21 0.18000000000000002])

product(F::Factor{T}...) -> Factor

Compute a factor product of an arbitrary number of factors.

JunctionTrees.marg — Function

marg(A::Factor{T, ND}, V::Tuple{Vararg{Int64, N}} where N) -> Factor

Sum out the variables in V from factor A.

Examples

A = Factor{Float64,2}((1, 2), [0.59 0.41; 0.22 0.78])
marg(A, (2,))

# output

Factor{Float64, 1}((1,), [1.0, 1.0])

marg(A::Factor, V::Int64...) -> Factor

Sum out an aribitrary number of variables from factor A.

Examples

A = Factor{Float64,3}((1, 2, 3), cat([0.25 0.08; 0.05 0.0; 0.15 0.09],
                                     [0.35 0.16; 0.07 0.0; 0.21 0.18], dims=3))
marg(A, 1, 2)

# output

Factor{Float64, 1}((3,), [0.6199999999999999, 0.97])

JunctionTrees.redu — Function

redu(A::Factor{T}, vars::Tuple, vals::Tuple) -> Factor

Reduce/invalidate all entries that are not consitent with the evidence.

Examples

A = Factor{Float64,3}((1, 2, 3), cat([0.25 0.08; 0.05 0.0; 0.15 0.09],
                                     [0.35 0.16; 0.07 0.0; 0.21 0.18], dims=3))
obs_vars = (3,)
obs_vals = (1,)
redu(A, obs_vars, obs_vals)

# output

Factor{Float64, 3}((1, 2, 3), [0.25 0.08; 0.05 0.0; 0.15 0.09;;; 0.0 0.0; 0.0 0.0; 0.0 0.0])

JunctionTrees.norm — Function

norm(A::Factor) -> Factor

Normalize the values in Factor A such that all probabilities lie between 0 and 1.

Examples

A = Factor{Float64,2}((1, 2), [0.2 0.4; 0.6 0.8])
norm(A)

# output

Factor{Float64, 2}((1, 2), [0.1 0.2; 0.3 0.4])

JunctionTrees.LastStage — Type

Enumerated type used to select up to which stage an expression of the junction tree algorithm should be returned.

JunctionTrees.ForwardPass — Constant

Return an expression up to and including the forward pass.

JunctionTrees.BackwardPass — Constant

Return an expression up to and including the backward pass.

JunctionTrees.JointMarginals — Constant

Return an expression that computes the cluster joint marginals.

JunctionTrees.UnnormalizedMarginals — Constant

Return an expression that computes the joint marginals.

JunctionTrees.Marginals — Constant

Return an expression that computes the posterior marginals.