Public Documentation

Documentation for JunctionTrees.jl's public interface.

See the Internals section of the manual for internal package docs.

Index

JunctionTrees

JunctionTrees.BackwardPass

JunctionTrees.ForwardPass

JunctionTrees.JointMarginals

JunctionTrees.Marginals

JunctionTrees.UnnormalizedMarginals

JunctionTrees.Factor

JunctionTrees.LastStage

JunctionTrees.computeMarginalsExpr

JunctionTrees.getGraph

JunctionTrees.marg

JunctionTrees.norm

JunctionTrees.product

JunctionTrees.redu

Public Interface

JunctionTrees — Module

Main module for JunctionTrees.jl – a Julia implementation of the junction tree algorithm.

One main function is exported from this module for public use:

computeMarginalsExpr. Computes the posterior marginals using the junction tree algorithm.

Exports

BackwardPass

Factor

ForwardPass

JointMarginals

LastStage

Marginals

UnnormalizedMarginals

computeMarginalsExpr

getGraph

marg

norm

product

redu

JunctionTrees.computeMarginalsExpr — Function

computeMarginalsExpr(uai_filepath::String; uai_evid_filepath, td_filepath, apply_partial_evaluation, last_stage, smart_root_selection) -> Expr

Return an expression containing the computations to compute the marginals of all the variables in the model using the junction tree algorithm.

JunctionTrees.Factor — Type

struct Factor{T, N}

vars::Tuple{Vararg{Int64, N}} where N

vals::Array

A composite type that implements the factor datatype.

JunctionTrees.product — Function

product(A::Factor{T}, B::Factor{T}) -> Factor

Compute a factor product of tables A and B.

Examples

A = Factor{Float64,2}((2, 1), [0.5 0.1 0.03; 0.8 0.0 0.9])
B = Factor{Float64,2}((3, 2), [0.5 0.1; 0.7 0.2])
C = product(A, B)

product(F::Factor{T}...) -> Factor

Compute a factor product of an arbitrary number of factor arguments.

JunctionTrees.marg — Function

marg(A::Factor{T, ND}, V::Tuple{Vararg{Int64, N}} where N) -> Factor

Sum out the variables in V from factor A.

Examples

A = Factor{Float64,3}((3, 2, 1), cat([0.25 0.08; 0.05 0.0; 0.15 0.09],
                                     [0.35 0.16; 0.07 0.0; 0.21 0.18], dims=3))
V = [1]
B = marg(A, V)

A = Factor{Float64,3}((2, 7, 1), cat([1 2; 3 4;  5  6],
                                     [7 8; 9 10; 11 12], dims=3))
V = [7]
B = marg(A, V)

marg(A::Factor, V::Int64...) -> Factor

Sum out an aribitrary number of variables from factor A.

JunctionTrees.redu — Function

redu(A::Factor{T}, vars::Tuple, vals::Tuple) -> Factor

Reduce/invalidate all entries that are not consitent with the evidence.

Examples

A = Factor{Float64,3}((1, 2, 3), cat([0.25 0.08; 0.05 0.0; 0.15 0.09],
                                     [0.35 0.16; 0.07 0.0; 0.21 0.18], dims=3))
vars = (1,3)
vals = (1,1)
B = redu(A, vars, vals)

JunctionTrees.norm — Function

norm(A::Factor) -> Factor

Normalize the values in Factor A such that all probabilities lie between 0 and 1.

Examples

A = Factor([2, 1], [1.0 2.0 3.0; 4.0 5.0 6.0])
B = norm(A)

JunctionTrees.getGraph — Function

getGraph() -> Any

Return the current graph.

JunctionTrees.LastStage — Type

Enumerated type used to select up to which stage an expression of the junction tree algorithm should be returned.

JunctionTrees.ForwardPass — Constant

Return an expression up to and including the forward pass.

JunctionTrees.BackwardPass — Constant

Return an expression up to and including the backward pass.

JunctionTrees.JointMarginals — Constant

Return an expression that computes the cluster joint marginals.

JunctionTrees.UnnormalizedMarginals — Constant

Return an expression that computes the joint marginals.

JunctionTrees.Marginals — Constant

Return an expression that computes the posterior marginals.